docs: improve inline code comments and formatting

+++ laracasts/laravel-api-master-class/app/Http/Controllers/Api/V1/UserController.php
@@ -1,5 +1,4 @@
 <?php
-
 namespace App\Http\Controllers\Api\V1;
 
 use App\Http\Filters\V1\AuthorFilter;
@@ -9,15 +8,20 @@ use App\Http\Requests\Api\V1\UpdateUserRequest;
 use App\Http\Resources\V1\UserResource;
 use App\Models\User;
 use App\Policies\V1\UserPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
-use Illuminate\Database\Eloquent\ModelNotFoundException;
 
 class UserController extends ApiController
 {
     protected $policyClass = UserPolicy::class;
 
     /**
-     * Display a listing of the resource.
+     * Get all users
+     *
+     * @group Managing Users
+     *
+     * @queryParam sort string Data field(s) to sort by. Separate multiple fields with commas. Denote descending sort with a minus sign. Example: sort=name
+     * @queryParam filter[name] Filter by status name. Wildcards are supported. No-example
+     * @queryParam filter[email] Filter by email. Wildcards are supported. No-example
+     *
      */
     public function index(AuthorFilter $filters)
     {
@@ -27,22 +31,27 @@ class UserController extends ApiController
     }
 
     /**
-     * Store a newly created resource in storage.
+     * Create a user
+     *
+     * @group Managing Users
+     *
+     * @response 200 {"data":{"type":"user","id":16,"attributes":{"name":"My User","email":"user@user.com","isManager":false},"links":{"self":"http:\/\/localhost:8000\/api\/v1\/authors\/16"}}}
      */
     public function store(StoreUserRequest $request)
     {
-        try {
-            // policy
-             $this->isAble('store', User::class);
+        if ($this->isAble('store', User::class)) {
+            return new UserResource(User::create($request->mappedAttributes()));
+        }
 
-             return new UserResource(User::create($request->mappedAttributes()));
-         } catch (AuthorizationException $ex) {
-             return $this->error('You are not authorized to create that resource', 401);
-         }
+        return $this->notAuthorized('You are not authorized to create that resource');
     }
 
     /**
-     * Display the specified resource.
+     * Display a user
+     *
+     * @group Managing Users
+     *
+     *
      */
     public function show(User $user)
     {
@@ -54,60 +63,58 @@ class UserController extends ApiController
     }
 
     /**
-     * Update the specified resource in storage.
+     * Update a user
+     *
+     * @group Managing Users
+     *
+     * @response 200 {"data":{"type":"user","id":16,"attributes":{"name":"My User","email":"user@user.com","isManager":false},"links":{"self":"http:\/\/localhost:8000\/api\/v1\/authors\/16"}}}
      */
-    public function update(UpdateUserRequest $request, $user_id)
+    public function update(UpdateUserRequest $request, User $user)
     {
-        try {
-            $user = User::findOrFail($user_id);
-
-            // policy
-            $this->isAble('update', $user);
-
+        if ($this->isAble('update', $user)) {
             $user->update($request->mappedAttributes());
 
             return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('Ticket cannot be found.', 404);
-        } catch (AuthorizationException $ex) {
-            return $this->error('You are not authorized to update that resource', 401);
         }
-    }
 
-    public function replace(ReplaceUserRequest $request, $user_id) {
-        // PUT
-        try {
-            $user = User::findOrFail($user_id);
-
-            // policy
-            $this->isAble('replace', $user);
-
-            $user->update($request->mappedAttributes());
-
-
-            return new UserResource($user);
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot be found.', 404);
-        }
+        return $this->notAuthorized('You are not authorized to update that resource');
     }
 
     /**
-     * Remove the specified resource from storage.
+     * Replace a user
+     *
+     * @group Managing Users
+     *
+     * @response 200 {"data":{"type":"user","id":16,"attributes":{"name":"My User","email":"user@user.com","isManager":false},"links":{"self":"http:\/\/localhost:8000\/api\/v1\/authors\/16"}}}
      */
-    public function destroy($user_id)
+    public function replace(ReplaceUserRequest $request, User $user)
     {
-        try {
-            $user = User::findOrFail($user_id);
+        // PUT
+        if ($this->isAble('replace', $user)) {
+            $user->update($request->mappedAttributes());
 
-            // policy
-            $this->isAble('delete', $user);
+            return new UserResource($user);
+        }
 
+        return $this->notAuthorized('You are not authorized to update that resource');
+    }
+
+    /**
+     * Delete a user
+     *
+     * @group Managing Users
+     *
+     * @response 200 {}
+     */
+    public function destroy(User $user)
+    {
+        if ($this->isAble('delete', $user)) {
             $user->delete();
 
             return $this->ok('User successfully deleted');
-        } catch (ModelNotFoundException $exception) {
-            return $this->error('User cannot found.', 404);
         }
+
+        return $this->notAuthorized('You are not authorized to delete that resource');
     }
 
 }
